Rename project to linksight and add catalog
