+++ poc/string_searcher.py
