+++ prototype/linksight/api/admin.py
@@ -1,4 +1,4 @@
 from django.contrib import admin
-from project_record_matching.api import models
+from linksight.api import models
 
 admin.site.register(models.Dataset)
