+++ prototype/linksight/api/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from project_record_matching.api import views
+from linksight.api import views
 
 urlpatterns = [
     path('datasets/', views.dataset_list),
