+++ prototype/linksight/app/catalog/index.html
@@ -1,0 +1,13 @@
+<!doctype html>
+<html>
+<head>
+  <meta charset="utf-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+  <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.png">
+  <title>Catalog</title>
+</head>
+<body>
+  <div id="catalog"></div>
+</body>
+</html>
