+++ prototype/linksight/app/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "catalog": "3.0.0-rc.4",
     "react": "16.3.1",
     "react-dom": "16.3.1",
     "react-filepond": "2.0.6",
@@ -12,6 +13,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "catalog-start": "catalog start",
+    "catalog-build": "catalog build"
   }
 }
