+++ prototype/linksight/app/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "project_record_matching",
-  "name": "project_record_matching",
+  "short_name": "linksight",
+  "name": "linksight",
   "icons": [
     {
       "src": "favicon.ico",
