+++ prototype/linksight/app/src/registerServiceWorker.js
