+++ prototype/linksight/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for project_record_matching project.
+Django settings for linksight project.
 
 Generated by 'django-admin startproject' using Django 2.0.5.
 
@@ -48,8 +48,8 @@ INSTALLED_APPS = [
     'rest_framework',
     'corsheaders',
 
-    'project_record_matching.api',
-    'project_record_matching.app',
+    'linksight.api',
+    'linksight.app',
 ]
 
 MIDDLEWARE = [
@@ -63,7 +63,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'project_record_matching.urls'
+ROOT_URLCONF = 'linksight.urls'
 
 TEMPLATES = [
     {
@@ -81,7 +81,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'project_record_matching.wsgi.application'
+WSGI_APPLICATION = 'linksight.wsgi.application'
 
 
 # Database
