+++ prototype/linksight/urls.py
@@ -4,6 +4,6 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path('api/', include('project_record_matching.api.urls')),
+    path('api/', include('linksight.api.urls')),
     path('admin/', admin.site.urls),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
