+++ prototype/linksight/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for project_record_matching project.
+WSGI config for linksight project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "project_record_matching.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "linksight.settings")
 
 application = get_wsgi_application()
